attached_assets/requirements.txt
telethon>=1.24.0
python-dotenv>=0.19.2
requests>=2.28.0
google-generativeai>=0.3.2
PySocks>=1.7.1
